+++ pages/modals/sdc_storage_stat.php
@@ -1,0 +1,88 @@
+<div class="modal fade" id="modal-abm-storage-stat">
+    <div class="modal-dialog">
+    <div class="modal-content">
+        <div class="modal-header">
+            <button type="button" class="close" data-dismiss="modal"
+                aria-label="Close">
+                <span aria-hidden="true">&times;</span></button>
+            <h2 class="modal-title" id='modal-abm-storage-stat-title'>Métricas</h2>
+        </div>
+        <div class="modal-body">
+            <!-- form start -->
+            <form method="post" role="form" action="">
+                
+                <div class="box-body">
+                    <table id="tbstat" class="display" width="100%">
+                        <thead>
+                            <tr>
+                            <th width="10%" align="center">Categoría</th>
+                            <th class="text-right">[TB] Capacidad Asignable</th>
+                            <th class="text-right">[TB] Asignado</th>
+                            <th class="text-right">[%] Asignado Actual</th>
+                            <th class="text-right">[TB] Disponible Estimado</th>
+                            </tr>
+                        </thead>
+                        <tbody>
+                            <?php
+                            $query = "SELECT 
+                                        cat.nombre as cat_nombre , 
+                                        SUM(capacidad_asignable) as capacidad_asignable, 
+                                        SUM(asignado_tb) as asignado_tb, 
+                                        SUM(asignado_tb*100/capacidad_asignable) as asignado_actual,
+                                        SUM(disponible_estimado) as disponible_estimado
+                                        FROM (
+                                        SELECT 
+                                            sto_raw.categoria, 
+                                            (sto_raw.capacidad_fisica_tb * sto_raw.per_asignacion_recomendado / 100) as capacidad_asignable, 
+                                            CASE WHEN (sto_raw.asignado_tb >= (sto_raw.capacidad_fisica_tb * sto_raw.per_asignacion_recomendado / 100)) THEN 0 
+                                                ELSE (sto_raw.capacidad_fisica_tb * sto_raw.per_asignacion_recomendado / 100) - sto_raw.asignado_tb END as disponible_estimado,
+                                            sto_raw.asignado_tb, 
+                                            sto_raw.capacidad_fisica_tb, 
+                                            sto_raw.per_asignacion_recomendado 
+                                        FROM sdc_storage as sto_raw
+                                            WHERE sto_raw.borrado = 0
+                                            AND sto_raw.estado = 1
+                                        ) as sto_gr
+                                        INNER JOIN sto_categorias as cat ON sto_gr.categoria = cat.id
+                                        GROUP BY cat_nombre;"; 
+                            
+                            $sql = mysqli_query($con, $query);
+
+                            if(mysqli_num_rows($sql) > 0){
+                                $no = 1;
+                                $sum_cap_asignable = 0;
+                                $sum_asignado = 0;
+                                $sum_disponible = 0;
+                                while($row = mysqli_fetch_assoc($sql)){
+                                    $sum_cap_asignable += $row['capacidad_asignable'];
+                                    $sum_asignado += $row['asignado_tb'];
+                                    $sum_disponible += $row['disponible_estimado'];
+                                    echo '<tr>';
+                                    echo '<td>'. $row['cat_nombre'].'</td>';
+                                    echo '<td class="text-right">'. number_format($row['capacidad_asignable'],2,",",".").'</td>';
+                                    echo '<td class="text-right">'. number_format($row['asignado_tb'],2,",",".").'</td>';
+                                    echo '<td class="text-right">'. setSemaphoreBadge($row['asignado_actual'], $_TIPO_RANGOS_ASIGNADOS, true).'</td>';
+                                    echo '<td class="text-right">'. setSemaphoreBadge($row['disponible_estimado'], $_TIPO_RANGOS_OCUPADOS, true).'</td>';
+                                    echo '</tr>';
+                                }
+                            }
+                            ?>
+                        </tbody>
+                        <tfoot>
+                            <?php
+                            echo '<th>Total</th>';
+                            echo '<th class="text-right">'. number_format($sum_cap_asignable,2,",",".") .'</th>';
+                            echo '<th class="text-right">'. number_format($sum_asignado,2,",",".") .'</th>';
+                            echo '<th></th>';
+                            echo '<th class="text-right">'. number_format($sum_disponible,2,",",".") .'</th>';
+                            ?>
+                        </tfoot>   
+                    </table>                    
+                </div>
+            </form>
+        </div>
+    </div>
+        <!-- /.modal-content -->
+    </div>
+    <!-- /.modal-dialog -->
+</div>